Docstrings

GeometricAlgebraModels.translate — Method

translate(p)
translate(p, X) -> X′ = sandwich_prod(translate(p), X)

Translation versor for the displacement vector p::Grade{1} for the algebra that X and p belong to.

The single-argument method returns the translation versor itself and the two-argument form applies the versor to X with GeometricAlgebra.sandwich_prod.

Different methods of translate(p::Grade{1,Sig}) may exist for different algebras Sig (for instance, it returns the conformal translation rotor if p is a CGA vector). The two-argument method promotes p and X into the same algebra before calling translate(p).

Conformal

GeometricAlgebraModels.Conformal — Module

Conformal geometric algebra

Tools for working with conformal geometric algebra (CGA).

The algebra CGA{Sig} is an extension of the base space Sig with two additional dimensions of square $+1$ and $-1$. Points in the base space p::Multivector{Sig,1} are associated to null vectors in the higher space by

up(x) = n0 + x + x^2/2*noo

where n0, noo = nullbasis() are the null vectors representing the origin and the point at infinity.

Geometric primitives including flats (points, point pairs, lines, planes, ...) and rounds (point pairs, circles, spheres, ...) can be naturally represented as the inner or outer product null spaces of CGA blades (see ipns and opns).

Geometric primitives may be translated in space with translate and combined with the wedge product to produce joins (for opns geometries) or meets/intersections (for ipns).

Any CGA blade has a standardform from which geometric properties (such as position, radius and carrier) can be easily retrieved.

GeometricAlgebraModels.Conformal.CGA — Type

Metric signature for the conformal geometric algebra over a base space with metric signature $Sig$. A conformal algebra CGA{Sig} has dimension dimension(Sig) + 2, with the two extra basis vectors squaring to $+1$ and $-1$, respectively.

GeometricAlgebraModels.Conformal.CGABlade — Type

CGABlade{Sig,K}:
	DirectionBlade{Sig,K}(E)
	FlatBlade{Sig,K}(E, p)
	DualFlatBlade{Sig,K}(E, p)
	RoundBlade{Sig,K}(E, p, r2)
	TangentBlade{Sig,K}(E, p)

Standard forms of blades in the conformal geometric algebra CGA{Sig} over base space Sig.

Value	Mathematical form

DirectionBlade(E)	$E ∧ n_∞$

FlatBlade(E, p)	$𝚃ₚ[n₀ ∧ E ∧ n_∞]$

DualFlatBlade(E, p)	$𝚃ₚ[E]$

RoundBlade(E, p, r2)	$𝚃ₚ[(n₀ + r2/2 n_∞) ∧ E]$

↳ TangentBlade(E, p)	$𝚃ₚ[n₀ ∧ E]$

Any blade in CGA{Sig} is of exactly one of the forms above, where:

$E$ is a K-blade in the base space

$p$ is a position vector in the base space

$r2$ is a square-radius, which may be positive or negative

$n₀$ and $n_∞$ are the points at the Conformal.origin and at Conformal.infinity

$Tₚ$ is the translation operator sending $n₀$ to $p$

The method standardform puts any blade in CGA{Sig} into one of these forms. A standard blade X::CGABlade may be converted back to the usual additive form with Multivector(X).

See table 14.1 of [1] for discussion.

GeometricAlgebraModels.Conformal.CGAGeometry — Type

CGAGeometry{Sig}:
	FlatGeometry{Sig,K}(p, E)
	RoundGeometry{Sig,K}(p, E, r2)
	PointAtInfinity{Sig}()
	EmptySet{Sig}()

Subsets of $ℝⁿ ∪ {∞}$ which are the ipns or opns of a conformal blade.

See also FlatGeometry and RoundGeometry.

GeometricAlgebraModels.Conformal.DirectionBlade — Type

CGABlade{Sig,K}:
	DirectionBlade{Sig,K}(E)
	FlatBlade{Sig,K}(E, p)
	DualFlatBlade{Sig,K}(E, p)
	RoundBlade{Sig,K}(E, p, r2)
	TangentBlade{Sig,K}(E, p)

Standard forms of blades in the conformal geometric algebra CGA{Sig} over base space Sig.

Value	Mathematical form

DirectionBlade(E)	$E ∧ n_∞$

FlatBlade(E, p)	$𝚃ₚ[n₀ ∧ E ∧ n_∞]$

DualFlatBlade(E, p)	$𝚃ₚ[E]$

RoundBlade(E, p, r2)	$𝚃ₚ[(n₀ + r2/2 n_∞) ∧ E]$

↳ TangentBlade(E, p)	$𝚃ₚ[n₀ ∧ E]$

Any blade in CGA{Sig} is of exactly one of the forms above, where:

$E$ is a K-blade in the base space

$p$ is a position vector in the base space

$r2$ is a square-radius, which may be positive or negative

$n₀$ and $n_∞$ are the points at the Conformal.origin and at Conformal.infinity

$Tₚ$ is the translation operator sending $n₀$ to $p$

The method standardform puts any blade in CGA{Sig} into one of these forms. A standard blade X::CGABlade may be converted back to the usual additive form with Multivector(X).

See table 14.1 of [1] for discussion.

GeometricAlgebraModels.Conformal.DualFlatBlade — Type

CGABlade{Sig,K}:
	DirectionBlade{Sig,K}(E)
	FlatBlade{Sig,K}(E, p)
	DualFlatBlade{Sig,K}(E, p)
	RoundBlade{Sig,K}(E, p, r2)
	TangentBlade{Sig,K}(E, p)

Standard forms of blades in the conformal geometric algebra CGA{Sig} over base space Sig.

Value	Mathematical form

DirectionBlade(E)	$E ∧ n_∞$

FlatBlade(E, p)	$𝚃ₚ[n₀ ∧ E ∧ n_∞]$

DualFlatBlade(E, p)	$𝚃ₚ[E]$

RoundBlade(E, p, r2)	$𝚃ₚ[(n₀ + r2/2 n_∞) ∧ E]$

↳ TangentBlade(E, p)	$𝚃ₚ[n₀ ∧ E]$

Any blade in CGA{Sig} is of exactly one of the forms above, where:

$E$ is a K-blade in the base space

$p$ is a position vector in the base space

$r2$ is a square-radius, which may be positive or negative

$n₀$ and $n_∞$ are the points at the Conformal.origin and at Conformal.infinity

$Tₚ$ is the translation operator sending $n₀$ to $p$

The method standardform puts any blade in CGA{Sig} into one of these forms. A standard blade X::CGABlade may be converted back to the usual additive form with Multivector(X).

See table 14.1 of [1] for discussion.

GeometricAlgebraModels.Conformal.EmptySet — Type

CGAGeometry{Sig}:
	FlatGeometry{Sig,K}(p, E)
	RoundGeometry{Sig,K}(p, E, r2)
	PointAtInfinity{Sig}()
	EmptySet{Sig}()

Subsets of $ℝⁿ ∪ {∞}$ which are the ipns or opns of a conformal blade.

See also FlatGeometry and RoundGeometry.

GeometricAlgebraModels.Conformal.FlatBlade — Type

CGABlade{Sig,K}:
	DirectionBlade{Sig,K}(E)
	FlatBlade{Sig,K}(E, p)
	DualFlatBlade{Sig,K}(E, p)
	RoundBlade{Sig,K}(E, p, r2)
	TangentBlade{Sig,K}(E, p)

Standard forms of blades in the conformal geometric algebra CGA{Sig} over base space Sig.

Value	Mathematical form

DirectionBlade(E)	$E ∧ n_∞$

FlatBlade(E, p)	$𝚃ₚ[n₀ ∧ E ∧ n_∞]$

DualFlatBlade(E, p)	$𝚃ₚ[E]$

RoundBlade(E, p, r2)	$𝚃ₚ[(n₀ + r2/2 n_∞) ∧ E]$

↳ TangentBlade(E, p)	$𝚃ₚ[n₀ ∧ E]$

Any blade in CGA{Sig} is of exactly one of the forms above, where:

$E$ is a K-blade in the base space

$p$ is a position vector in the base space

$r2$ is a square-radius, which may be positive or negative

$n₀$ and $n_∞$ are the points at the Conformal.origin and at Conformal.infinity

$Tₚ$ is the translation operator sending $n₀$ to $p$

The method standardform puts any blade in CGA{Sig} into one of these forms. A standard blade X::CGABlade may be converted back to the usual additive form with Multivector(X).

See table 14.1 of [1] for discussion.

GeometricAlgebraModels.Conformal.FlatGeometry — Type

FlatGeometry{Sig,K}(p, E) <: CGAGeometry{Sig}

A K-flat in the base space Sig through the point p::Multivector{Sig,1} spanning the K-blade E::Multivector{Sig,K}.

All flats include the unique point at infinity.

$k$-flat	name

$0$-flat	point

$1$-flat	line

$2$-flat	plane

$3$-flat	volume

See also RoundGeometry and CGAGeometry.

GeometricAlgebraModels.Conformal.PointAtInfinity — Type

CGAGeometry{Sig}:
	FlatGeometry{Sig,K}(p, E)
	RoundGeometry{Sig,K}(p, E, r2)
	PointAtInfinity{Sig}()
	EmptySet{Sig}()

Subsets of $ℝⁿ ∪ {∞}$ which are the ipns or opns of a conformal blade.

See also FlatGeometry and RoundGeometry.

GeometricAlgebraModels.Conformal.RoundBlade — Type

CGABlade{Sig,K}:
	DirectionBlade{Sig,K}(E)
	FlatBlade{Sig,K}(E, p)
	DualFlatBlade{Sig,K}(E, p)
	RoundBlade{Sig,K}(E, p, r2)
	TangentBlade{Sig,K}(E, p)

Standard forms of blades in the conformal geometric algebra CGA{Sig} over base space Sig.

Value	Mathematical form

DirectionBlade(E)	$E ∧ n_∞$

FlatBlade(E, p)	$𝚃ₚ[n₀ ∧ E ∧ n_∞]$

DualFlatBlade(E, p)	$𝚃ₚ[E]$

RoundBlade(E, p, r2)	$𝚃ₚ[(n₀ + r2/2 n_∞) ∧ E]$

↳ TangentBlade(E, p)	$𝚃ₚ[n₀ ∧ E]$

Any blade in CGA{Sig} is of exactly one of the forms above, where:

$E$ is a K-blade in the base space

$p$ is a position vector in the base space

$r2$ is a square-radius, which may be positive or negative

$n₀$ and $n_∞$ are the points at the Conformal.origin and at Conformal.infinity

$Tₚ$ is the translation operator sending $n₀$ to $p$

The method standardform puts any blade in CGA{Sig} into one of these forms. A standard blade X::CGABlade may be converted back to the usual additive form with Multivector(X).

See table 14.1 of [1] for discussion.

GeometricAlgebraModels.Conformal.RoundGeometry — Type

RoundGeometry{Sig,K}(p, E, r2) <: CGAGeometry{Sig}

A K-round in the base space Sig with center point p::Multivector{Sig,1} spanning the carrier K-blade E::Multivector{Sig,K} with square radius r2.

A $k$-round is a $(k - 1)$-sphere, whose carrier is a $k$-plane. No rounds contain the unique point at infinity.

$k$-round	$(k - 1)$-sphere	name

$0$-round	$-1$-sphere	empty set

$1$-round	$0$-sphere	point pair

$2$-round	$1$-sphere	circle

$3$-round	$2$-sphere	sphere

Note

The square radius r2 may be negative, in which case the round is formally the empty set, but one may also interpret this as an "imaginary" radius.

See also FlatGeometry and CGAGeometry.

GeometricAlgebraModels.Conformal.TangentBlade — Type

TangentBlade{Sig,K}

A TangentBlade is a special case of RoundBlade which has zero square radius.

Value	Mathematical form

RoundBlade(E, p, r2)	$𝚃ₚ[(n₀ + r2/2 n_∞) ∧ E]$

TangentBlade(E)	$𝚃ₚ[n₀ ∧ E]$

GeometricAlgebraModels.Conformal.infinity — Function

nullbasis(n) = (origin(n), infinity(n))
origin(n)
infinity(n)

Standard null basis vectors in the conformal geometric algebra CGA{n}.

The point at the origin $e₀$ and the point at infinity $e∞$ in the n-dimensional conformal geometric algebra model are defined as $e₀ = (e₊ + e₋)/2$ and $e∞ = e₋ - e₊$ where $e₊$ and $e₋$ are the standard extra basis vectors squaring to $+1$ and $-1$.

Type-stable methods exist which accept the type CGA{n} instead of an integer n.

GeometricAlgebraModels.Conformal.ipns — Function

ipns(A) -> CGAGeometry
opns(A) -> CGAGeometry

Inner or outer product null space of the blade A.

This is the set of points $x$ satisfying up(x)⋅A ≈ 0 (IPNS) or up(x)∧A ≈ 0 (OPNS), possibly including the point at infinity.

GeometricAlgebraModels.Conformal.nullbasis — Function

nullbasis(n) = (origin(n), infinity(n))
origin(n)
infinity(n)

Standard null basis vectors in the conformal geometric algebra CGA{n}.

The point at the origin $e₀$ and the point at infinity $e∞$ in the n-dimensional conformal geometric algebra model are defined as $e₀ = (e₊ + e₋)/2$ and $e∞ = e₋ - e₊$ where $e₊$ and $e₋$ are the standard extra basis vectors squaring to $+1$ and $-1$.

Type-stable methods exist which accept the type CGA{n} instead of an integer n.

GeometricAlgebraModels.Conformal.opns — Function

ipns(A) -> CGAGeometry
opns(A) -> CGAGeometry

Inner or outer product null space of the blade A.

This is the set of points $x$ satisfying up(x)⋅A ≈ 0 (IPNS) or up(x)∧A ≈ 0 (OPNS), possibly including the point at infinity.

GeometricAlgebraModels.Conformal.origin — Function

nullbasis(n) = (origin(n), infinity(n))
origin(n)
infinity(n)

Standard null basis vectors in the conformal geometric algebra CGA{n}.

The point at the origin $e₀$ and the point at infinity $e∞$ in the n-dimensional conformal geometric algebra model are defined as $e₀ = (e₊ + e₋)/2$ and $e∞ = e₋ - e₊$ where $e₊$ and $e₋$ are the standard extra basis vectors squaring to $+1$ and $-1$.

Type-stable methods exist which accept the type CGA{n} instead of an integer n.

GeometricAlgebraModels.Conformal.up — Function

up(::Multivector{Sig,1}) -> Multivector{CGA{Sig},1}
dn(::Multivector{CGA{Sig},1}) -> Multivector{Sig,1}

Conformal embedding of a point.

Lift "up" a 1-vector in the base space Sig to a null vector in the conformal algebra CGA{Sig}, or project "down" a conformal 1-vector back into the base space.

The up map is given by

\[up(x) = n0 + embed(x) + 1/2 x^2 noo\]

where n0, noo = nullbasis(CGA{Sig}) are the points representing the origin and infinity.

For any vector u we have dn(up(u)) == n and up∘dn is idempotent.

GeometricAlgebraModels.dn — Function

up(::Multivector{Sig,1}) -> Multivector{CGA{Sig},1}
dn(::Multivector{CGA{Sig},1}) -> Multivector{Sig,1}

Conformal embedding of a point.

Lift "up" a 1-vector in the base space Sig to a null vector in the conformal algebra CGA{Sig}, or project "down" a conformal 1-vector back into the base space.

The up map is given by

\[up(x) = n0 + embed(x) + 1/2 x^2 noo\]

where n0, noo = nullbasis(CGA{Sig}) are the points representing the origin and infinity.

For any vector u we have dn(up(u)) == n and up∘dn is idempotent.

GeometricAlgebraModels.standardform — Method

standardform(X::AbstractMultivector{CGA{Sig}}) -> CGABlade{Sig}

Put the blade X in standard form, returning a CGABlade object.

GeometricAlgebraModels.translate — Method

translate(p::Grade{1,CGA{Sig}})

Translation versor for conformal geometric algebra.

Examples

julia> p, x = randn(Multivector{3,1}, 2);

julia> translate(p, up(x)) ≈ up(x + p)
true

julia> translate(p, infinity()) ≈ infinity()
true

1Dorst, L., Fontijne, D., & Mann, S. (2010). Geometric Algebra for Computer Science: An Object-Oriented Approach to Geometry. Elsevier.

1Dorst, L., Fontijne, D., & Mann, S. (2010). Geometric Algebra for Computer Science: An Object-Oriented Approach to Geometry. Elsevier.

1Dorst, L., Fontijne, D., & Mann, S. (2010). Geometric Algebra for Computer Science: An Object-Oriented Approach to Geometry. Elsevier.

1Dorst, L., Fontijne, D., & Mann, S. (2010). Geometric Algebra for Computer Science: An Object-Oriented Approach to Geometry. Elsevier.

1Dorst, L., Fontijne, D., & Mann, S. (2010). Geometric Algebra for Computer Science: An Object-Oriented Approach to Geometry. Elsevier.